backend/app/tenant/jwt_validator.py
"""
Clerk JWT-Validierung via JWKS (MiroFish)

Clerk signiert JWTs mit RS256. Die öffentlichen Schlüssel werden vom
JWKS-Endpoint abgerufen und 5 Minuten gecacht (PyJWKClient mit lifespan).

Relevante JWT-Claims:
    sub:      Clerk User ID
    org_id:   Clerk Organization ID  → wird für Tenant-Lookup verwendet
    org_slug: Organization Slug
    exp:      Ablaufzeitpunkt
"""

import threading
from typing import Any, Dict

import jwt
from jwt import InvalidTokenError, PyJWKClient  # PyJWT >= 2.8.0
from app.utils.logger import get_logger

logger = get_logger("mirofish.tenant.jwt")

# Ein PyJWKClient pro JWKS-URL (Thread-sicher gecacht)
_clients: Dict[str, PyJWKClient] = {}
_lock = threading.Lock()


def _get_client(jwks_url: str) -> PyJWKClient:
    """Gibt einen gecachten PyJWKClient für die URL zurück."""
    if jwks_url not in _clients:
        with _lock:
            if jwks_url not in _clients:
                _clients[jwks_url] = PyJWKClient(
                    jwks_url,
                    cache_jwk_set=True,
                    lifespan=300,  # 5 Minuten JWKS-Cache
                )
                logger.info(f"JWKS-Client initialisiert: {jwks_url}")
    return _clients[jwks_url]


def validate_clerk_token(token: str, jwks_url: str) -> Dict[str, Any]:
    """
    Validiert ein Clerk JWT und gibt den Payload zurück.

    Args:
        token:    Bearer-Token aus dem Authorization-Header
        jwks_url: URL zum JWKS-Endpoint

    Returns:
        Decodierter JWT-Payload

    Raises:
        jwt.ExpiredSignatureError:  Token ist abgelaufen
        jwt.InvalidTokenError:      Token ist ungültig
    """
    client = _get_client(jwks_url)

    try:
        signing_key = client.get_signing_key_from_jwt(token)
    except Exception as exc:
        raise InvalidTokenError(f"JWKS-Schlüssel nicht gefunden: {exc}") from exc

    return jwt.decode(
        token,
        signing_key.key,
        algorithms=["RS256"],
        options={"require": ["exp", "sub"]},
        leeway=30,  # 30 Sekunden Toleranz für Uhrzeit-Abweichungen
    )
